several files were relocated to reflect the project requirements
